🍱 Separate express server from index script to avoid force exit on jest
